Use IJetSelector tool interface instead of using JetCleaningTool directly (TrigT1CaloMonitoring-00-14-13)
    * Use IJetSelector tool interface instead of using JetCleaningTool directly
    * Tag TrigT1CaloMonitoring-00-14-13

2014-11-10  Sasha Mazurov <alexander.mazurov@cern.ch
	* Fix Jet tool for using JetCleaningTool instead of deprecated methods
	  from JetUtils
	* Fix compilation warnings.
	* Tag TrigT1CaloMonitoring-00-14-12

2014-10-07  Hanno Meyer zu Theenhausen <hanno.meyer.zu.theenhausen@cern.ch>
	* Tag as TrigT1CaloMonitoring-00-14-11
	* src/PPrMon.cxx, TrigT1CaloMonitoring/PPrMon.h
	 - Add monitoring for LUT-JEP with dummy values
	 - Rename LUT to LUT-CP

+++ Trigger/TrigT1/TrigT1CaloMonitoring/src/JEMMon.cxx
@@ -303,7 +303,7 @@ StatusCode JEMMon::bookHistogramsRecurrent()
     // This code is back compativle with RUN1 (l1version==0)
     // More on theresholds configuration see:
     //    TrigConf::L1DataDef::setMaxThresholdsFromL1Versionl1version
-    for (int i = 0; i < jetNames.size(); i++) {
+    for (size_t i = 0; i < jetNames.size(); i++) {
       buffer.str("");
       buffer << i;
       name  = "jem_2d_etaPhi_roi_MainThresh" + buffer.str();
